StudentLanguagesDetailController moved to new folder Student
	renamed:    app/Http/Controllers/api/StudentLanguagesDetailController.php -> app/Http/Controllers/api/Student/StudentLanguagesDetailController.php
	modified:   routes/api/v1.php
	modified:   tests/Feature/Student/StudentLanguagesDetailControllerTest.php
