+++ app/Http/Controllers/api/Student/StudentLanguagesDetailController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Controllers\api;
+namespace App\Http\Controllers\api\Student;
 
 use Exception;
 use App\Http\Controllers\Controller;
